refactor: extract session timeline component
